Allow publish jfrog-cli-go npm package with Jenkins

+++ Jenkinsfile
@@ -7,30 +7,34 @@ node {
         [pkg: 'jfrog-cli-windows-amd64', goos: 'windows', goarch: 'amd64', fileExtention: '.exe']
     ]
 
-    subject = 'jfrog'
-    repo = 'jfrog-cli-go'
-    sh 'rm -rf temp'
-    sh 'mkdir temp'
-    def goRoot = tool 'go-1.7.1'
-    dir('temp'){
-        cliWorkspace = pwd()
-        withEnv(["GOROOT=$goRoot","GOPATH=${cliWorkspace}","PATH+GOROOT=${goRoot}/bin", "JFROG_CLI_OFFER_CONFIG=false"]) {
+    if ("$PUBLISH_NPM_PACKAGE") {
+        publishNpmPackage()
+    } else {
+        subject = 'jfrog'
+        repo = 'jfrog-cli-go'
+        sh 'rm -rf temp'
+        sh 'mkdir temp'
+        def goRoot = tool 'go-1.7.1'
+        dir('temp'){
+            cliWorkspace = pwd()
+            withEnv(["GOROOT=$goRoot","GOPATH=${cliWorkspace}","PATH+GOROOT=${goRoot}/bin", "JFROG_CLI_OFFER_CONFIG=false"]) {
 
-            stage 'Go get'
-            sh 'go get -f -u github.com/jfrogdev/jfrog-cli-go/jfrog-cli/jfrog'
-            if (BRANCH?.trim()) {
-                dir("src/github.com/jfrogdev/jfrog-cli-go/jfrog-cli/jfrog") {
-                    sh "git checkout $BRANCH"
-                    sh 'go install'
+                stage 'Go get'
+                sh 'go get -f -u github.com/jfrogdev/jfrog-cli-go/jfrog-cli/jfrog'
+                if (BRANCH?.trim()) {
+                    dir("src/github.com/jfrogdev/jfrog-cli-go/jfrog-cli/jfrog") {
+                        sh "git checkout $BRANCH"
+                        sh 'go install'
+                    }
                 }
-            }
-            sh 'bin/jfrog --version > version'
-            version = readFile('version').trim().split(" ")[2]
-            print "publishing version: $version"
-            for (int i = 0; i < architectures.size(); i++) {
-                    def currentBuild = architectures[i]
-                    stage "Build ${currentBuild.pkg}"
-                    buildAndUpload(currentBuild.goos, currentBuild.goarch, currentBuild.pkg, currentBuild.fileExtention)
+                sh 'bin/jfrog --version > version'
+                version = readFile('version').trim().split(" ")[2]
+                print "publishing version: $version"
+                for (int i = 0; i < architectures.size(); i++) {
+                        def currentBuild = architectures[i]
+                        stage "Build ${currentBuild.pkg}"
+                        buildAndUpload(currentBuild.goos, currentBuild.goarch, currentBuild.pkg, currentBuild.fileExtention)
+                }
             }
         }
     }
@@ -47,4 +51,11 @@ def buildAndUpload(goos, goarch, pkg, fileExtension) {
     def extension = fileExtension == null ? '' : fileExtension
     uploadToBintray(pkg, extension)
     sh "rm jfrog$extension"
+}
+
+def publishNpmPackage() {
+    dir ('jfrog-cli-go/npm/') {
+        sh "echo "//registry.npmjs.org/:_authToken=$NPM_AUTH_TOKEN""
+        sh "npm publish"
+    }
 }
